Split plainly-plugin into src/ui and src/node folders
