+++ plainly-aescripts/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "clean": "rm -rf dist",
     "compile": "yarn clean && tsc",
-    "dev": "concurrently \"tsc --watch\" \"webpack --mode development --watch\"",
+    "dev": "yarn compile && concurrently \"tsc --watch\" \"webpack --mode development --watch\"",
     "build-test": "yarn compile && webpack --mode production",
     "build": "yarn compile && webpack --mode production",
     "biome": "biome check ./src",
