+++ plainly-plugin/src/ui/components/logo/PlainlyLogo.tsx
