+++ plainly-plugin/src/ui/components/navigation/Sidebar.tsx
