+++ plainly-plugin/src/ui/components/navigation/SidebarResources.tsx
